requirements.txt
gradio
requests
pandas==2.2.3
langchain-core
langchain-openai
langgraph==0.4.3
python-dotenv

beautifulsoup4==4.13.4
langchain-community
wolframalpha==5.1.3
wikipedia==1.4.0

# Check if useful for your project
playwright
pydantic==2.11.3
